+++ xrGame/MathUtils.h
@@ -279,7 +279,7 @@ IC void twoq_2w(const Fquaternion &q1, const Fquaternion &q2, float dt, Fvector 
 	w.add(v1);
 	float sinus_2 = 1.f - cosinus * cosinus, k = 2.f / dt;
 	if (sinus_2 > EPS)
-		k *= acos(cosinus) / _sqrt(sinus_2);
+		k *= acosf(cosinus) / _sqrt(sinus_2);
 	w.mul(k);
 }
