+++ xrGame/alife_simulator_base.cpp
@@ -92,7 +92,7 @@ CSE_Abstract *CALifeSimulatorBase::spawn_item(LPCSTR section, const Fvector &pos
 	R_ASSERT3(abstract, "Cannot find item with section", section);
 
 	abstract->s_name = section;
-	abstract->s_gameid = server().game->Type();
+	abstract->s_gameid = static_cast<u8>(server().game->Type());
 	abstract->s_RP = 0xff;
 	abstract->ID = server().PerformIDgen(0xffff);
 	abstract->ID_Parent = parent_id;
