+++ xrGame/ik/eqn.cxx
@@ -63,7 +63,7 @@ static int solve_trig1_aux(float c,
 	if (temp < 0.0f)
 		return 0;
 
-	temp = atan2(_sqrt(temp), c);
+	temp = atan2f(_sqrt(temp), c);
 	num = (_abs(temp) > 1e-6f) ? 2 : 1;
 
 	theta[0] = atan2ba;
@@ -94,7 +94,7 @@ static int solve_trig1_aux(float c,
 
 static int solve_trig1(float a, float b, float c, float theta[2])
 {
-	return solve_trig1_aux(c, a * a + b * b, atan2(b, a), theta);
+	return solve_trig1_aux(c, a * a + b * b, atan2f(b, a), theta);
 }
 
 #if 0
