+++ xrGame/ui/UIMMShniaga.cpp
@@ -367,7 +367,7 @@ float CUIMMShniaga::pos(float x1, float x2, u32 t)
 	float x = 0;
 
 	if (t >= 0 && t <= m_run_time)
-		x = log(1 + (t * 10.0f) / m_run_time) / log(11.0f);
+		x = logf(1 + (t * 10.0f) / m_run_time) / logf(11.0f);
 	else if (t <= 0)
 		x = 0;
 	else if (t > m_run_time)
